refactor(Catch Log Statistics): extract utils (LAN-875)
