+++ landa/water_body_management/report/catch_log_statistics/catch_log_statistics.py
@@ -1,24 +1,21 @@
 # Copyright (c) 2022, Real Experts GmbH and contributors
 # For license information, please see license.txt
 
-from typing import List
+from typing import TYPE_CHECKING, List
 
 import frappe
 from frappe import _
 from pypika.functions import Coalesce, Substring, Sum
-from pypika.queries import Table
-from pypika.terms import Criterion
 
-from landa.organization_management.doctype.organization.organization import (
-	get_supported_water_bodies,
+from landa.water_body_management.report.utils import (
+	add_conditions,
+	add_or_filters,
+	is_regional_or_state_employee,
 )
-from landa.utils import get_current_member_data
 
-STATE_ROLES = {"LANDA State Organization Employee", "System Manager", "Administrator"}
-REGIONAL_ROLES = {
-	"LANDA Regional Organization Management",
-	"LANDA Regional Water Body Management",
-}
+if TYPE_CHECKING:
+	from pypika.queries import Table
+	from pypika.terms import Criterion
 
 
 def get_columns(
@@ -228,7 +225,7 @@ def get_data(
 	return query.run()
 
 
-def get_subquery(entry: Table, child_table: Table, qb_filters: List[Criterion]):
+def get_subquery(entry: "Table", child_table: "Table", qb_filters: "List[Criterion]"):
 	subquery = (
 		frappe.qb.from_(entry)
 		.join(child_table)
@@ -244,7 +241,11 @@ def get_subquery(entry: Table, child_table: Table, qb_filters: List[Criterion]):
 
 
 def filter_and_group(
-	query, entry: Table, child_table: Table, qb_filters: List[Criterion], group_by_fish_species=False
+	query,
+	entry: "Table",
+	child_table: "Table",
+	qb_filters: "List[Criterion]",
+	group_by_fish_species=False,
 ):
 	query = add_conditions(query, qb_filters)
 	query = add_or_filters(query, entry)
@@ -282,58 +283,6 @@ def get_qb_filters(filters, entry, child_table):
 		qb_filters.append(entry.year <= to_year)
 
 	return qb_filters
-
-
-def add_or_filters(query, entry):
-	"""Return a dict of filters that restricts the results to what the user is
-	allowed to see.
-
-	STATE_ROLES		no filters
-	REGIONAL_ROLES	everything related to their water bodys OR to their member organizations
-	LOCAL_ROLES		everything related to their own organization and OR to the water bodys it is supporting
-	"""
-	user_roles = get_user_roles()
-
-	if user_roles.intersection(STATE_ROLES):
-		return query
-
-	# User is not a state organization employee
-
-	member_data = get_current_member_data()
-	if not member_data:
-		frappe.throw(_("You are not a member of any organization."))
-
-	if user_roles.intersection(REGIONAL_ROLES):
-		return query.where(
-			(entry.regional_organization == member_data.regional_organization)
-			| entry.organization.like(f"{member_data.regional_organization}-%")
-		)
-
-	# User is not in regional organization management
-	supported_water_bodies = get_supported_water_bodies(member_data.local_organization)
-	if supported_water_bodies:
-		return query.where(
-			entry.organization.like(f"{member_data.local_organization}%")
-			| entry.water_body.isin(supported_water_bodies)
-		)
-
-	return query.where(entry.organization.like(f"{member_data.local_organization}%"))
-
-
-def add_conditions(query, conditions):
-	for condition in conditions:
-		query = query.where(condition)
-
-	return query
-
-
-def get_user_roles():
-	return set(frappe.get_roles())
-
-
-def is_regional_or_state_employee():
-	user_roles = get_user_roles()
-	return REGIONAL_ROLES.intersection(user_roles) or STATE_ROLES.intersection(user_roles)
 
 
 def execute(filters=None):
